+++ app/src/main/java/com/github/droidworksstudio/launcher/ui/bottomsheetdialog/PaddingBottomSheetDialogFragment.kt
@@ -9,6 +9,7 @@ import android.view.ViewGroup
 import androidx.annotation.RequiresApi
 import androidx.fragment.app.viewModels
 import com.github.droidworksstudio.launcher.databinding.BottomsheetdialogPaddingSettingsBinding
+import com.github.droidworksstudio.launcher.helper.AppHelper
 import com.github.droidworksstudio.launcher.helper.BottomDialogHelper
 import com.github.droidworksstudio.launcher.helper.PreferenceHelper
 import com.github.droidworksstudio.launcher.viewmodel.PreferenceViewModel
@@ -26,6 +27,9 @@ class PaddingBottomSheetDialogFragment : BottomSheetDialogFragment() {
     lateinit var preferenceHelper: PreferenceHelper
 
     @Inject
+    lateinit var appHelper: AppHelper
+
+    @Inject
     lateinit var bottomDialogHelper: BottomDialogHelper
 
     private val preferenceViewModel: PreferenceViewModel by viewModels()
@@ -38,7 +42,7 @@ class PaddingBottomSheetDialogFragment : BottomSheetDialogFragment() {
         return binding.root
     }
 
-    @RequiresApi(Build.VERSION_CODES.O)
+    @RequiresApi(Build.VERSION_CODES.Q)
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -56,6 +60,7 @@ class PaddingBottomSheetDialogFragment : BottomSheetDialogFragment() {
         binding.selectAppPaddingSize.setText("${preferenceHelper.homeAppPadding}")
     }
 
+    @RequiresApi(Build.VERSION_CODES.Q)
     private fun observeValueChange() {
         val appValue = binding.selectAppPaddingSize.text.toString()
 
@@ -63,6 +68,8 @@ class PaddingBottomSheetDialogFragment : BottomSheetDialogFragment() {
         dismiss()
 
         preferenceViewModel.setAppPaddingSize(appFloatValue)
+        val feedbackType = "select"
+        appHelper.triggerHapticFeedback(context, feedbackType)
     }
 
     private fun parseFloatValue(text: String, defaultValue: Float): Float {
@@ -72,6 +79,7 @@ class PaddingBottomSheetDialogFragment : BottomSheetDialogFragment() {
         return text.toFloat()
     }
 
+    @RequiresApi(Build.VERSION_CODES.Q)
     override fun onDismiss(dialog: DialogInterface) {
         super.onDismiss(dialog)