+++ app/src/main/java/com/github/droidworksstudio/launcher/ui/home/HomeFragment.kt
@@ -79,7 +79,6 @@ class HomeFragment : Fragment(),
     @Inject
     lateinit var appHelper: AppHelper
 
-
     @Inject
     lateinit var fingerHelper: BiometricHelper